requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.37
psycopg[binary,async]==3.2.4
sentence-transformers==3.4.0
transformers==4.48.1
pillow==11.1.0
python-dotenv==1.0.1
accelerate==1.3.0
passlib==1.7.4
requests==2.32.3
pydantic-settings==2.7.1
pgvector==0.3.6
asyncpg==0.30.0
# torch==2.5.1